Catch strings naming sockets and skip reading them
The logic here looks for strings that look like files and attempts to
read them. If we can determine that the string points to an AF_UNIX
socket, we can assume we shouldn't try to read it.

For now, we'll assume that the best guess as to the scope path is the
same as when the user specifies the connection as a hostname or
hostname:port - the current directory or java_root.